[DataCollection] fixed an upload bug of images.
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@40587 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/DataCollection/classes/class.ilDataCollectionRecord.php
@@ -542,8 +542,10 @@ class ilDataCollectionRecord
 	 */
 	public function deleteFile($obj_id)
 	{
-		$file = new ilObjFile($obj_id, false);
-		$file->delete();
+        if(ilObject2::_lookupObjId($obj_id)){
+		    $file = new ilObjFile($obj_id, false);
+		    $file->delete();
+        }
 	}
